introduce static wrapper class
