+++ Src/MoneyFox.Ui/Views/Statistics/CategorySummary/StatisticCategorySummaryViewModel.cs
@@ -41,7 +41,7 @@ internal sealed class StatisticCategorySummaryViewModel : StatisticViewModel
     {
         try
         {
-            var categorySummaryModel = await Mediator.Send(new GetCategorySummaryQuery { EndDate = EndDate, StartDate = StartDate });
+            var categorySummaryModel = await Mediator.Send(new GetCategorySummary.Query(StartDate: EndDate, EndDate: StartDate));
             CategorySummary = new(
                 categorySummaryModel.CategoryOverviewItems.Select(
                     x => new CategoryOverviewViewModel
